Added FileTransfer Session Start and Session Stop messages. These were implemented to  allow the server to know when the file transfer GUI is in use on the remote system. User impersonation is not released until the file transfer session ends. This allows the file transfer session to continue until the user dismisses the file transfer window.
A new rfb pseudo encoding was added to allow the viewer to query the server for the FT protocol version that it supports.

Refactored viewer code to use new m_ServerFTProtocolVersion variable instead of m_fOldProtocole. This is more future-proof.

Fixed typo: buzy is now busy
Fixed a couple of compile errors from bad merges.

git-svn-id: https://svn.code.sf.net/p/ultravnc/code/UltraVNC%20Project%20Root/UltraVNC@311 0bd26501-c230-0410-ab9e-a7946180fe3b