Fix broken screen color with vnc4server in 32 bit color depth.
If UltraVNC viewer connects to vnc4server (in Debian jessie), with
32 bit color depth, blue component is completeley lost. This is
caused by bugs of both vnc4server and UltraVNC viewer. This patch
fixes the bug in UltraVNC side. (However, due to the bug in
vnc4server side, viewer fails to connect to vnc4server with 32 bit
color depth even with this patch.)