Fix connection issue with vnc4server in 32 bit color depth.
UltraVNC viewer fails to connect with vnc4server in 32 bit color
depth due to a bug in vnc4server side. If the comment-out of this
patch is removed, the patch resolves the issue by fallback the
connection to 24 bit color depth. (With this patch, the codes added
are commented out by default.)